Clear styles before printing the username in the prompt
Sometimes, especially after running `l`, the user in the prompt would
not be styled in **bold**; when clearing the screen however, it would be
styled just fine.

Anyway, I figured there was something _weird_ with `tree`'s output (`l`
is a wrapper around `tree`), so I tried to reset every style before my
custom prompt and that seems to have fixed the problem.