Fix prettier-indentexpr sometimes returning a wrong indent level
Running prettier against the current buffer might (and most likely will)
change the total number of lines of the current buffer:

- long lines will be broken into smaller ones
- unnecessarily short lines will be joined into longer ones

This means that we cannot expect to find the current line (i.e. the line
at a:lnum) to be stable; instead, we have to add a marker to the current
line, run prettier, and then find that marker again and see how many
levels did it get indented with.

It's not perfect, and it's not super efficient, but it works most of the
times, so I can live with it.